+++ ReleaseNotes.md
@@ -2,6 +2,22 @@
 
 Solace Event Portal SDK.
 
+## Version 0.8.0-alpha
+
+**Enhancements:**
+* **VersionsServices**
+  - added paging when retrieving latest version to:
+    - EpSdkEnumVersionsService
+    - EpSdkEpEventVersionsService
+    - EpSdkEventApiVersionsService
+    - EpSdkSchemaVersionsService
+
+**Tests:**
+- added test coverage using nyc
+
+**License**
+- moved to apache 2.0
+
 ## Version 0.7.0-alpha
 
 **Documentation:**